+++ Common/src/main/java/ram/talia/hexal/common/entities/IMediaEntity.kt
@@ -3,13 +3,13 @@ package ram.talia.hexal.common.entities
 import com.mojang.datafixers.util.Either
 import net.minecraft.server.level.ServerPlayer
 import net.minecraft.world.entity.Entity
-import java.lang.Integer.max
+import kotlin.math.max
 
 interface IMediaEntity<T : Entity> {
 	/**
 	 * Should always be >= 0.
 	 */
-	var media: Int
+	var media: Long
 
 	val isConsumable: Boolean
 
@@ -21,9 +21,9 @@ interface IMediaEntity<T : Entity> {
 
 	fun get(): T
 
-	fun addMedia(dMedia: Int) {
+	fun addMedia(dMedia: Long) {
 		val new = media + dMedia
 		// preventing overflow above Int.MAX_VALUE and disallowing values below 0.
-		media = if (dMedia > 0 && media > new) Int.MAX_VALUE - 1 else max(new, 0)
+		media = if (dMedia > 0 && media > new) Long.MAX_VALUE - 1 else max(new, 0)
 	}
 }
